Refactor: break test_views in a package
